misc: paintbox: Make bulk allocation blocking
Adding blocking and timeout behavior to the
bulk_allocation ioctl.

Caller to the ioctl can specify timeout by setting
the timeout_ns field in the input struct ipu_bulk_allocation_request

Setting the timeout_ns to 0 will cause the allocation ioctl
to be non-blocking, that is, it will return if the resources
requested are not available.

Setting the timeout_ns to LONG_MAX will cause the allocation
ioctl to wait infinitely for the resources to become available.

Setting the timeout_ns to other value will cause the allocation
to block until either the resource becomes available, or it has
timed out, whichever comes first.

Change-Id: Ic09e670a2a20cddc5758c98a61114734d57e177d
Signed-off-by: Alex Tang <pztang@google.com>